+++ server/svix-server/src/v1/mod.rs
@@ -14,8 +14,7 @@ pub fn router() -> Router {
         .merge(endpoints::endpoint::router())
         .merge(endpoints::event_type::router())
         .merge(endpoints::message::router())
-        .merge(endpoints::attempt::router())
-        .merge(endpoints::retry_schedule::router());
+        .merge(endpoints::attempt::router());
 
     #[cfg(debug_assertions)]
     if cfg!(debug_assertions) {
